# Add fp8 per block quant triton kernel