E2E Speed up for Qwen3-30B-A3B-FP8

From 22 seconds to 20 seconds, we should see more gain for larger model like 671B and 235B 



### Here is the detailed profiling result:

#### Before: It took 397us for the tensor quantization.
<img width="973" height="615" alt="Screenshot 2025-07-27 at 12 06 14 PM" src="https://github.com/user-attachments/assets/8625ce33-4f9e-44ad-862a-d736d74fcc28" />

#### After: It took 173us for the tensor quantization
<img width="975" height="636" alt="Screenshot 2025-07-27 at 12 06 58 PM" src="https://github.com/user-attachments/assets/745b07fa-611a-485c-9b46-be6c2aab7f6e" />